Enables adding products to the cart
Adds functionality to add products to the cart from the product list and product details pages.
This feature sends a request to the backend to add the selected product to the user's cart.
It also includes an alert to notify the user when a product is successfully added to the cart.